requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
